fundoo/requirement.txt
aioredis==1.2.0
amqp==2.5.0
apns==2.0.1
apturl==0.5.2
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
awscli==1.14.44
bcrypt==3.1.6
billiard==3.6.0.0
boto==2.49.0
boto3==1.9.245
botocore==1.12.245
Brlapi==0.6.6
cached-property==1.5.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
CodeConvert==3.0.2
colorama==0.3.7
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.7
cupshelpers==1.0
daphne==2.3.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
django-admin==2.0.0
django-allauth==0.40.0
django-cors-headers==3.1.1
django-excel-response2==2.0.8
django-model-utils==3.2.0
django-models-ext==1.1.9
django-redis==4.10.0
django-rest-swagger==2.1.2
django-restframework==0.0.1
django-security==0.11.1
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.7.2
django-storages-redux==1.3.3
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.1.1
furl==2.1.0
get==2019.4.13
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
httpie==0.9.8
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.22
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.3
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
linecache2==1.0.0
locust==0.0
locustio==0.11.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.1
msgpack-python==0.5.6
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.4
netifaces==0.10.4
numpy==1.16.3
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.1
paramiko==2.5.0
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
pluggy==0.13.0
post==2019.4.13
powerline-status==2.6
protobuf==3.0.0
psutil==5.6.3
public==2019.4.13
py==1.8.0
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyee==6.0.0
pyfcm==1.4.7
Pygments==2.2.0
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pytest==5.1.2
python-apt==1.6.4
python-crontab==2.3.8
python-dateutil==2.8.0
python-debian==0.1.32
python-decouple==3.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
query-string==2019.4.13
redis==2.10.6
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.2.1
screen==1.0.1
SecretStorage==2.3.1
shortuuid==0.5.0
simplejson==3.13.2
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
texttable==0.9.1
TimeConvert==1.5.0
traceback2==1.4.0
Twisted==19.2.0
txaio==18.8.1
typed-ast==1.4.0
tzlocal==2.0.0
ua-parser==0.8.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
usb-creator==0.3.3
validate-email==1.3
vine==1.3.0
virtualenv==16.6.0
virtualenv-clone==0.5.3
wadllib==1.3.2
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
xkit==0.0.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.6.0
